requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
langchain==0.0.339
chromadb==0.4.18
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
openai==1.3.5
tiktoken==0.5.1
pandas==2.1.3
numpy==1.26.2
loguru==0.7.2
scrapy==2.11.0
selenium==4.15.2
fastapi-utils==0.2.1 